requirements.txt
requests==2.31.0
python-telegram-bot==20.3
python-dateutil==2.8.2
pytz==2024.1
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
upstash-redis==0.15.0
